+++ src/pulumi/core-dtap/analytics/databricks/workspaces/engineering.py
@@ -1,5 +1,6 @@
 import os
 import pulumi
+from pulumi.resource import ResourceOptions
 import pulumi_azure_native as azure_native
 import pulumi_azuread as azuread
 
@@ -13,12 +14,12 @@ from ingenii_azure_data_platform.iam import (
 )
 from ingenii_azure_data_platform.utils import generate_hash, generate_resource_name
 
-from project_config import azure_client, platform_config, platform_outputs
+from project_config import azure_client, platform_config, platform_outputs, DTAP_ROOT
 from management import resource_groups
 from management.user_groups import user_groups
 from security import credentials_store
 from network import vnet
-from storage.datalake import datalake
+from storage.datalake import datalake, datalake_containers
 
 outputs = platform_outputs["analytics"]["databricks"]["workspaces"]["engineering"] = {}
 
@@ -73,14 +74,10 @@ outputs["url"] = workspace.workspace_url
 # ----------------------------------------------------------------------------------------------------------------------
 # ENGINEERING DATABRICKS WORKSPACE -> IAM ROLE ASSIGNMENTS
 # ----------------------------------------------------------------------------------------------------------------------
-try:
-    iam_role_assignments = workspace_config["iam"]["role_assignments"]
-except:
-    iam_role_assignments = {}
 
 # TODO: Create a function that takes care of the role assignments. Replace all role assignments using the function.
 # Create role assignments defined in the YAML files
-for assignment in iam_role_assignments:
+for assignment in workspace_config.get("iam", {}).get("role_assignments", []):
     # User Group Assignment
     user_group_ref_key = assignment.get("user_group_ref_key")
     if user_group_ref_key is not None:
@@ -166,6 +163,22 @@ datafactory_token = databricks.Token(
 )
 
 # ----------------------------------------------------------------------------------------------------------------------
+# ENGINEERING DATABRICKS WORKSPACE -> PRE-PROCESSING PACKAGE
+# ----------------------------------------------------------------------------------------------------------------------
+
+blob_name = "pre_process-1.0.0-py3-none-any.whl"
+pre_processing_package = pulumi.FileAsset(f"{DTAP_ROOT}/assets/{blob_name}")
+pre_processing_blob = azure_native.storage.Blob(
+    resource_name=f'{workspace_name}-pre_processing_package',
+    account_name=datalake.name,
+    blob_name="pre_process/" + blob_name,
+    container_name=datalake_containers["utilities"].name,
+    resource_group_name=resource_groups["data"].name,
+    source=pre_processing_package,
+    opts=pulumi.ResourceOptions(ignore_changes=["content_md5"]),
+)
+
+# ----------------------------------------------------------------------------------------------------------------------
 # ENGINEERING DATABRICKS WORKSPACE -> CLUSTERS
 # ----------------------------------------------------------------------------------------------------------------------
 
@@ -181,16 +194,25 @@ for ref_key, cluster_config in cluster_definitions.items():
     )
 
     # Cluster Libraries
+    libraries = cluster_config.get("libraries", {})
     pypi_libraries = [
         databricks.ClusterLibraryArgs(
             pypi=databricks.ClusterLibraryPypiArgs(
                 package=lib.get("package"), repo=lib.get("repo")
             )
         )
-        for lib in cluster_config.get("libraries", {}).get("pypi", [])
+        for lib in libraries.get("pypi", [])
+    ]
+    whl_libraries = [
+        databricks.ClusterLibraryArgs(
+            whl="dbfs:/mnt/utilities/pre_process/pre_process-1.0.0-py3-none-any.whl"
+        )
+    ] + [
+        databricks.ClusterLibraryArgs(whl=lib)
+        for lib in libraries.get("whl", [])
     ]
 
-    cluster_libraries = [*pypi_libraries]
+    cluster_libraries = pypi_libraries + whl_libraries
 
     base_cluster_configuration = {
         "resource_name": cluster_resource_name,
@@ -210,7 +232,10 @@ for ref_key, cluster_config in cluster_definitions.items():
             "DBT_LOGS_FOLDER": "/dbfs/mnt/dbt-logs",
             **cluster_config.get("spark_env_vars", {}),
         },
-        "opts": pulumi.ResourceOptions(provider=databricks_provider),
+        "opts": pulumi.ResourceOptions(
+            provider=databricks_provider, 
+            depends_on=[pre_processing_blob]
+        ),
     }
     if cluster_config.get("docker_image_url"):
         base_cluster_configuration["docker_image"] = databricks.ClusterDockerImageArgs(